Conflicts:
	autotest

+++ eoxserver/services/ows/wps/parameters/complexdata.py
@@ -183,10 +183,11 @@ class CDFile(CDBase):
         self._remove_file = remove_file
 
     def __del__(self):
-        name = self.name
-        self.close()
-        if self._remove_file:
-            os.remove(name)
+        if hasattr(self, "_file"):
+            name = self.name
+            self.close()
+            if self._remove_file:
+                os.remove(name)
 
     @property
     def data(self):
@@ -194,7 +195,11 @@ class CDFile(CDBase):
         return self.read()
 
     def __getattr__(self, attr):
-        return getattr(self._file, attr)
+        if attr == "_file":
+            raise AttributeError("Instance has no attribute '_file'")
+        else:
+            # Allow object to behave like a file.
+            return getattr(self._file, attr)
 
 
 class CDPermanentFile(CDFile):
